casts removed

+++ src/processes/Input.java
@@ -35,8 +35,8 @@ public class Input extends Process {
 		
 		
 		Kernel.getResources().destroy("inputstart");
-		RealMachineRegisters.setCH2((byte) 1);
-		RealMachineRegisters.setIOI((byte) 0);
+		RealMachineRegisters.setCH2(1);
+		RealMachineRegisters.setIOI(0);
 	}
 
 }
